requirements.txt
torch>=2.1.0
numpy>=1.21.0
gymnasium>=1.2.0
matplotlib>=3.7.0
pyyaml>=6.0
